Add ability to archive paster.xyz
You can now download an archive of the paster database in the form of a
gzipped tar file. The name of each file is the ID of the paste in raw
URL base64 encoding and the modified date is the creation date of the
paste. The content of each file is obviously the paste itself.